Add explicit cast to uint16_t for rx count
Suppress implicit narrowing conversion warning by explicitly casting the DMA transfer count difference to uint16_t.

+++ lib-gd32/src/uart0/uart0.cpp
@@ -56,7 +56,7 @@ extern "C" void USART0_IRQHandler() {
     if (RESET != usart_interrupt_flag_get(USART0, USART_INT_FLAG_IDLE)) {
         usart_data_receive(USART0);
 
-        sv_rx_count.rx = kSizeRxBuffer - (dma_transfer_number_get(USART0_DMAx, USART0_RX_DMA_CHx));
+        sv_rx_count.rx = static_cast<uint16_t>(kSizeRxBuffer - (dma_transfer_number_get(USART0_DMAx, USART0_RX_DMA_CHx)));
         sv_receive_flag = 1;
 
         auto chtl = DMA_CHCTL(USART0_DMAx, USART0_RX_DMA_CHx);
